+++ waspc/e2e-test/test-outputs/waspCompile-golden/waspCompile/.wasp/out/web-app/src/router.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Route, BrowserRouter as Router } from 'react-router-dom'
+import { Route, Switch, BrowserRouter as Router } from 'react-router-dom'
 
 
 import MainPage from './ext-src/MainPage.jsx'
@@ -7,9 +7,9 @@ import MainPage from './ext-src/MainPage.jsx'
 
 const router = (
   <Router>
-    <>
+    <Switch>
       <Route exact path="/" component={ MainPage }/>
-    </>
+    </Switch>
   </Router>
 )