+++ waspc/ChangeLog.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes
 - Wasp CLI will now forward error exit codes. This will help when used in scripted contexts.
+- Wasp now renders only the first route that matches the current path in the browser. 
 
 ### Express middleware customization
 We now offer the ability to customize Express middleware: